renamed absence reducer and created membersReducer
